Asset archive cooker working, add detailed cook progress output, add cooked archive viewer to imgui editor, improve file mapping to use long instead of int

+++ XRENGINE/Core/Files/FileMap.cs
@@ -9,12 +9,12 @@ namespace System
     public abstract class FileMap : IDisposable
     {
         protected VoidPtr _addr;
-        protected int _length;
+        protected long _length;
         protected string? _path;
         protected FileStream? _baseStream;
 
         public VoidPtr Address => _addr;
-        public int Length { get => _length; set => _length = value; }
+        public long Length { get => _length; set => _length = value; }
         public string? FilePath => _path;
         public FileStream? BaseStream => _baseStream;
 
@@ -34,9 +34,9 @@ namespace System
             => FromFile(path, FileMapProtect.ReadWrite, 0, 0);
         public static FileMap FromFile(string path, FileMapProtect prot) 
             => FromFile(path, prot, 0, 0);
-        public static FileMap FromFile(string path, FileMapProtect prot, int offset, int length)
+        public static FileMap FromFile(string path, FileMapProtect prot, long offset, long length)
             => FromFile(path, prot, offset, length, FileOptions.RandomAccess);
-        public static FileMap FromFile(string path, FileMapProtect prot, int offset, int length, FileOptions options)
+        public static FileMap FromFile(string path, FileMapProtect prot, long offset, long length, FileOptions options)
         {
             FileStream stream;
             FileMap map;
@@ -66,9 +66,9 @@ namespace System
             map._path = path; //In case we're using a temp file
             return map;
         }
-        public static FileMap? FromTempFile(int length)
+        public static FileMap? FromTempFile(long length)
             => FromTempFile(length, out _);
-        public static FileMap? FromTempFile(int length, out string path)
+        public static FileMap? FromTempFile(long length, out string path)
         {
             FileStream stream = new(path = Path.GetTempFileName(), FileMode.Open, FileAccess.ReadWrite, FileShare.Read, 8, FileOptions.RandomAccess | FileOptions.DeleteOnClose);
             try
@@ -87,36 +87,32 @@ namespace System
             => FromStream(stream, FileMapProtect.ReadWrite, 0, 0);
         public static FileMap FromStream(FileStream stream, FileMapProtect prot)
             => FromStream(stream, prot, 0, 0);
-        public static FileMap FromStream(FileStream stream, FileMapProtect prot, int offset, int length)
+        public static FileMap FromStream(FileStream stream, FileMapProtect prot, long offset, long length)
         {
-            //FileStream newStream = new FileStream(stream.Name, FileMode.Open, prot == FileMapProtect.Read ? FileAccess.Read : FileAccess.ReadWrite, FileShare.Read, 8, FileOptions.RandomAccess);
-            //try { return FromStreamInternal(newStream, prot, offset, length); }
-            //catch (Exception x) { newStream.Dispose(); throw x; }
-
             if (length == 0)
-                length = (int)stream.Length;
+                length = stream.Length;
             else
-                length = length.ClampMax((int)stream.Length);
+                length = length.ClampMax(stream.Length);
 
             return Environment.OSVersion.Platform switch
             {
-                PlatformID.Win32NT => new WFileMap(stream.SafeFileHandle.DangerousGetHandle(), prot, offset, (uint)length) { _path = stream.Name },
+                PlatformID.Win32NT => new WFileMap(stream.SafeFileHandle.DangerousGetHandle(), prot, offset, length) { _path = stream.Name },
                 _ => new CFileMap(stream, prot, offset, length) { _path = stream.Name },
             };
         }
 
-        public static FileMap FromStreamInternal(FileStream stream, FileMapProtect prot, int offset, int length)
+        public static FileMap FromStreamInternal(FileStream stream, FileMapProtect prot, long offset, long length)
         {
             if (length == 0)
-                length = (int)stream.Length;
+                length = stream.Length;
             else
-                length = length.ClampMax((int)stream.Length);
+                length = length.ClampMax(stream.Length);
 
-            length = length.ClampMin((int)stream.Length);
+            length = length.ClampMin(stream.Length);
 
             return Environment.OSVersion.Platform switch
             {
-                PlatformID.Win32NT => new WFileMap(stream.SafeFileHandle.DangerousGetHandle(), prot, offset, (uint)length) { _baseStream = stream, _path = stream.Name },
+                PlatformID.Win32NT => new WFileMap(stream.SafeFileHandle.DangerousGetHandle(), prot, offset, length) { _baseStream = stream, _path = stream.Name },
                 _ => new CFileMap(stream, prot, offset, length) { _baseStream = stream, _path = stream.Name },
             };
         }
@@ -130,7 +126,7 @@ namespace System
 
     public class WFileMap : FileMap
     {
-        internal WFileMap(VoidPtr hFile, FileMapProtect protect, long offset, uint length)
+        internal WFileMap(VoidPtr hFile, FileMapProtect protect, long offset, long length)
         {
             long maxSize = offset + length;
             uint maxHigh = (uint)(maxSize >> 32);
@@ -150,10 +146,10 @@ namespace System
 
             using Win32.SafeHandle h = Win32.CreateFileMapping(hFile, null, mProtect, maxHigh, maxLow, string.Empty);
             h.ErrorCheck();
-            _addr = Win32.MapViewOfFile(h.Handle, mAccess, (uint)(offset >> 32), (uint)offset, length);
+            _addr = Win32.MapViewOfFile(h.Handle, mAccess, (uint)(offset >> 32), (uint)offset, (nuint)length);
             if (!_addr)
                 Marshal.ThrowExceptionForHR(Marshal.GetHRForLastWin32Error());
-            _length = (int)length;
+            _length = length;
         }
 
         public override void Dispose()
@@ -173,7 +169,7 @@ namespace System
         protected MemoryMappedFile _mappedFile;
         protected MemoryMappedViewAccessor _mappedFileAccessor;
 
-        public CFileMap(FileStream stream, FileMapProtect protect, int offset, int length)
+        public CFileMap(FileStream stream, FileMapProtect protect, long offset, long length)
         {
             MemoryMappedFileAccess cProtect = (protect == FileMapProtect.ReadWrite) ? MemoryMappedFileAccess.ReadWrite : MemoryMappedFileAccess.Read;
             _length = length;
